+++ packages/react-chat/src/common/index.ts
@@ -1,3 +1,2 @@
-export * from './hooks';
 export * from './types';
 export * from './utils';
