+++ packages/react-chat/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAutoScroll';
 export * from './useDidUpdateEffect';
 export * from './useStateRef';
+export * from './useTheme';
