+++ packages/react-chat/src/hooks/useTheme.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 
 import { createCustomTheme } from '@/styles';
 
-import { Assistant } from './types';
+import { Assistant } from '../common/types';
 
 export const useTheme = (assistant?: Assistant) => {
   const [theme, setTheme] = useState('');
